optimize hash lookups
We used to do two map lookups on each Get(), one for the LRU, and one
for the SLRU.  This combines both maps into a single map so we can
figure out which list a node is on with a single lookup.

The lru.get() and slru.get() methods now just perform the list
rearrangements required for a successful get for those caches.

This also reduces the number of siphash invocations slightly as it will
be cached for all items in the slru cache.

Finally, for a number of routines which took or returned a key/value
pair, those are now updated to return an slruItem struct.